+++ gearoenix/render/widget/rnd-wdg-text.hpp
@@ -1,6 +1,7 @@
 #ifndef GEAROENIX_RENDER_WIDGET_TEXT_HPP
 #define GEAROENIX_RENDER_WIDGET_TEXT_HPP
 #include "../../core/cr-static.hpp"
+#include "../../math/math-vector-4d.hpp"
 #include "rnd-wdg-alignment.hpp"
 #include "rnd-wdg-widget.hpp"
 
@@ -12,7 +13,7 @@ namespace gearoenix::render::widget {
 class Text : public Widget {
     GX_GET_CREF_PRT(std::wstring, text)
     GX_GET_CREF_PRT(std::shared_ptr<font::Font2D>, text_font)
-    GX_GET_CREF_PRT(math::Vec4<float>, text_color)
+    GX_GET_CREF_PRT(math::Vec4<double>, text_color)
     GX_GET_VAL_PRT(Alignment, v_align, Alignment::Center)
     GX_GET_VAL_PRT(Alignment, h_align, Alignment::Center)
     GX_GET_VAL_PRT(core::Id, text_mesh_id, 0)
@@ -31,10 +32,10 @@ public:
         const std::wstring& t,
         const core::sync::EndCaller<core::sync::EndCallerIgnore>& c = GX_DEFAULT_IGNORED_END_CALLER) noexcept;
     void set_text_color(
-        float red, float green, float blue,
+        double red, double green, double blue,
         const core::sync::EndCaller<core::sync::EndCallerIgnore>& c = GX_DEFAULT_IGNORED_END_CALLER) noexcept;
     void set_text_color(
-        const math::Vec4<float>& v,
+        const math::Vec4<double>& v,
         const core::sync::EndCaller<core::sync::EndCallerIgnore>& c = GX_DEFAULT_IGNORED_END_CALLER) noexcept;
 };
 }
